tests/wpt/web-platform-tests/css/css-multicol/parsing/column-width-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: parsing column-width with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-width">
<meta name="assert" content="column-width supports only the grammar 'auto | <length>'.">
<meta name="assert" content="column-width rejects negative lengths.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("column-width", "none");
test_invalid_value("column-width", "10");
test_invalid_value("column-width", "-20px");
test_invalid_value("column-width", "30%");

test_invalid_value("column-width", "auto 40px");
</script>
</body>
</html>
